// src/api/models/StartSurveyDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SurveyId } from './SurveyId';
export type StartSurveyDto = {
  surveyId: SurveyId;
  /**
   * When true, the user opts out of participating in the survey.
   */
  optOut: boolean;
};